change package version format to ignore changes in this repository
The previous format made the package screen look a little confusing,
as all Chawan versions were there, but the most recent was at the bottom,
and using git short hashes would technically force people to go look up
the right version from Chawan git log. This process is clearer, but also
adds a small risk that a change in this repository that affects the AppImage
in some material way gets a skipped build until upstream changes something.
We will ignore this risk, and assume that I delete the package manually to
reproduce it in the event this happens.